requirements.txt
# Core Dependencies
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
requests==2.31.0
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.1.4
openpyxl==3.1.2

# Web Framework
streamlit==1.28.0
streamlit-chat==0.1.1
plotly==5.18.0

# ML & AI
torch==2.1.2
transformers==4.36.2
numpy==1.24.3
scikit-learn==1.3.2

# LLM Integration (Optional)
openai==1.6.1
anthropic==0.8.1

# Utilities
tqdm==4.66.1
pydantic==2.5.3
pyyaml==6.0.1
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
ipython==8.19.0
jupyter==1.0.0
